Cleanups, minor bugfixes, and add todos
Fix bug where the last note of an enemy could not be deleted.
Add convenience methods to delete RIV effects and Traits; needed to fix
bug where Traits and RIV effects could not be updated.
Add TODOs relative to changing the Enemy struct.
Change "remove" and derivates to "delete", for consistency.